factor out parseViolationRecord and test it
