+++ src/components/charts/new_revocations/CaseTable.js
@@ -20,6 +20,7 @@ import React, { useState, useEffect, useRef } from 'react';
 import ExportMenu from '../ExportMenu';
 
 import { COLORS } from '../../../assets/scripts/constants/colors';
+import parseViolationRecord from '../../../utils/charts/parseViolationRecord';
 import {
   getTrailingLabelFromMetricPeriodMonthsToggle, getPeriodLabelFromMetricPeriodMonthsToggle,
 } from '../../../utils/charts/toggles';
@@ -28,9 +29,6 @@ import {
 } from '../../../utils/transforms/labels';
 
 const CASES_PER_PAGE = 15;
-const VIOLATION_SEVERITY = [
-  'fel', 'misd', 'absc', 'muni', 'subs', 'tech',
-];
 
 const unknownStyle = {
   fontStyle: 'italic',
@@ -105,32 +103,6 @@ const CaseTable = (props) => {
       return <td>{label}</td>;
     }
     return <td style={unknownStyle}>{nullSafeLabel(label)}</td>;
-  };
-
-  const indexOf = (element, array) => {
-    for (let i = 0; i < array.length; i += 1) {
-      if (element === array[i]) {
-        return i;
-      }
-    }
-    return -1;
-  };
-
-  const parseViolationRecord = (recordLabel) => {
-    if (!recordLabel) {
-      return '';
-    }
-
-    const recordParts = recordLabel.split(';');
-    const records = recordParts.map((recordPart) => {
-      const number = recordPart[0];
-      const abbreviation = recordPart.substring(1);
-      return { number, abbreviation };
-    });
-    records.sort((a, b) => indexOf(a.abbreviation, VIOLATION_SEVERITY)
-      - indexOf(b.abbreviation, VIOLATION_SEVERITY));
-
-    return records.map((record) => `${record.number} ${record.abbreviation}`).join(', ');
   };
 
   return (
